refactor equipment package to adt package
