+++ src/com/dd/entities/adt/ItemType.java
@@ -1,4 +1,4 @@
-package com.dd.entities.equipments;
+package com.dd.entities.adt;
 
 public enum ItemType {
 	ARTIFACT, MAGICAL, ONEHANDEDWEAPON, POTION, SHIELD, SUIT, TWOHANDEDWEAPON, NONE;
